Refactor url/songId conversion
